moved all legacy files in pretri and renamed appropriately
