+++ _common/gameModel/pages/5.json
@@ -21,7 +21,7 @@
                             "style" : { },
                             "text" : {
                               "@class" : "Script",
-                              "content" : "import { registerSetStateAndThrottle } from \"./game/legacy/TimeManager\";\nregisterSetStateAndThrottle(Context.worldTime.setState);\n\"\"",
+                              "content" : "import { registerSetStateAndThrottle } from \"./game/pretri/pretriTime\";\nregisterSetStateAndThrottle(Context.worldTime.setState);\n\"\"",
                               "language" : "typescript"
                             }
                           },
@@ -37,7 +37,7 @@
                             "style" : { },
                             "text" : {
                               "@class" : "Script",
-                              "content" : "import { getWorldTime } from './game/legacy/TimeManager';\n\"Simulation Time: \" + getWorldTime().time;",
+                              "content" : "import { getWorldTime } from './game/pretri/pretriTime';\n\"Simulation Time: \" + getWorldTime().time;",
                               "language" : "JavaScript"
                             }
                           },
@@ -74,7 +74,7 @@
                             "localScriptEval" : {
                               "script" : {
                                 "@class" : "Script",
-                                "content" : "import { clearState } from './game/legacy/the_world';\n\nclearState();",
+                                "content" : "import { clearState } from './game/pretri/patientProcessing';\n\nclearState();",
                                 "language" : "TypeScript"
                               }
                             },
@@ -97,7 +97,7 @@
                         "id" : "",
                         "initialSelectedPageId" : {
                           "@class" : "Script",
-                          "content" : "import { getGamePageId } from './game/legacy/gameMaster';\ngetGamePageId()",
+                          "content" : "import { getGamePageId } from './game/pretri/drillMode';\ngetGamePageId()",
                           "language" : "TypeScript"
                         },
                         "layoutClassName" : "noPadding",
@@ -150,7 +150,7 @@
                         "className" : "paused-overlay noPadding noMargin",
                         "hideIf" : {
                           "@class" : "Script",
-                          "content" : "import { isDrillMode } from \"./game/legacy/gameMaster\";\n!isDrillMode() ||  Variable.find(gameModel, 'running_global').getValue(self);",
+                          "content" : "import { isDrillMode } from \"./game/pretri/drillMode\";\n!isDrillMode() ||  Variable.find(gameModel, 'running_global').getValue(self);",
                           "language" : "typescript"
                         },
                         "layout" : {
@@ -210,7 +210,7 @@
                 "id" : "",
                 "initialSelectedPageId" : {
                   "@class" : "Script",
-                  "content" : "import { getGamePageId } from './game/legacy/gameMaster';\ngetGamePageId()",
+                  "content" : "import { getGamePageId } from './game/pretri/drillMode';\ngetGamePageId()",
                   "language" : "TypeScript"
                 },
                 "layoutClassName" : "",
@@ -233,7 +233,7 @@
         "id" : "ifID",
         "ifCondition" : {
           "@class" : "Script",
-          "content" : "import { getTimeMode } from './game/legacy/gameMaster';\ngetTimeMode() === 'LIVE_WORLD';",
+          "content" : "import { getTimeMode } from './game/pretri/drillMode';\ngetTimeMode() === 'LIVE_WORLD';",
           "language" : "JavaScript"
         },
         "layoutClassName" : "noPadding noMargin",
