+++ basic_scenario/gameModel/gamemodel.json
@@ -6336,7 +6336,7 @@
   }, {
     "@class" : "GameModelContent",
     "content" : "",
-    "contentKey" : "game/legacy/gameMaster",
+    "contentKey" : "game/pretri/drillMode",
     "contentType" : "application/typescript",
     "libraryType" : "ClientScript",
     "refId" : "GameModelContent:#73:Nu8q8h",
@@ -6345,7 +6345,7 @@
   }, {
     "@class" : "GameModelContent",
     "content" : "",
-    "contentKey" : "game/legacy/likert",
+    "contentKey" : "game/pretri/likert",
     "contentType" : "application/typescript",
     "libraryType" : "ClientScript",
     "refId" : "GameModelContent:#67:7aiTeC",
@@ -6354,7 +6354,7 @@
   }, {
     "@class" : "GameModelContent",
     "content" : "",
-    "contentKey" : "game/legacy/the_world",
+    "contentKey" : "game/pretri/patientProcessing",
     "contentType" : "application/typescript",
     "libraryType" : "ClientScript",
     "refId" : "GameModelContent:#68:h6AnTo",
@@ -6363,7 +6363,7 @@
   }, {
     "@class" : "GameModelContent",
     "content" : "",
-    "contentKey" : "game/legacy/TimeManager",
+    "contentKey" : "game/pretri/pretriTime",
     "contentType" : "application/typescript",
     "libraryType" : "ClientScript",
     "refId" : "GameModelContent:#72:gtpcJn",
