+++ empty_scenario/gameModel/gamemodel.json
@@ -6251,7 +6251,7 @@
   }, {
     "@class" : "GameModelContent",
     "content" : "",
-    "contentKey" : "game/legacy/gameMaster",
+    "contentKey" : "game/pretri/drillMode",
     "contentType" : "application/typescript",
     "libraryType" : "ClientScript",
     "refId" : "GameModelContent:#73:Nu8q8h",
@@ -6260,7 +6260,7 @@
   }, {
     "@class" : "GameModelContent",
     "content" : "",
-    "contentKey" : "game/legacy/likert",
+    "contentKey" : "game/pretri/likert",
     "contentType" : "application/typescript",
     "libraryType" : "ClientScript",
     "refId" : "GameModelContent:#67:7aiTeC",
@@ -6269,7 +6269,7 @@
   }, {
     "@class" : "GameModelContent",
     "content" : "",
-    "contentKey" : "game/legacy/the_world",
+    "contentKey" : "game/pretri/patientProcessing",
     "contentType" : "application/typescript",
     "libraryType" : "ClientScript",
     "refId" : "GameModelContent:#68:h6AnTo",
@@ -6278,7 +6278,7 @@
   }, {
     "@class" : "GameModelContent",
     "content" : "",
-    "contentKey" : "game/legacy/TimeManager",
+    "contentKey" : "game/pretri/pretriTime",
     "contentType" : "application/typescript",
     "libraryType" : "ClientScript",
     "refId" : "GameModelContent:#72:gtpcJn",
